Move endpoints.health.mapping to management.health.status.http-mapping
Closes gh-10052
