+++ spring-boot-actuator/src/main/java/org/springframework/boot/actuate/autoconfigure/health/HealthIndicatorProperties.java
@@ -16,7 +16,9 @@
 
 package org.springframework.boot.actuate.autoconfigure.health;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
@@ -34,6 +36,12 @@ public class HealthIndicatorProperties {
 	 */
 	private List<String> order = null;
 
+	/**
+	 * Mapping of health statuses to HttpStatus codes. By default, registered health
+	 * statuses map to sensible defaults (i.e. UP maps to 200).
+	 */
+	private final Map<String, Integer> httpMapping = new HashMap<>();
+
 	public List<String> getOrder() {
 		return this.order;
 	}
@@ -42,6 +50,10 @@ public class HealthIndicatorProperties {
 		if (statusOrder != null && !statusOrder.isEmpty()) {
 			this.order = statusOrder;
 		}
+	}
+
+	public Map<String, Integer> getHttpMapping() {
+		return this.httpMapping;
 	}
 
 }
